update filenames & references and reorganize
