+++ index.html
@@ -3,7 +3,7 @@
 
 <head>
   <title>Share Button Test</title>
-  <link href="dist/styles.min.css" rel="stylesheet">
+  <link href="dist/ShareButtonStyles.min.css" rel="stylesheet">
   <style>
     html {
       width: 100%;
