+++ tests/fixtures/whatsapp.html
@@ -3,7 +3,7 @@
 <head>
   <title>Share Button Test: Whatsapp</title>
   <!-- temporary link tag -->
-  <link rel="stylesheet" type="text/css" href="../../dist/styles.min.css">
+  <link rel="stylesheet" type="text/css" href="../../dist/ShareButtonStyles.min.css">
   <style>
     html {
       width: 100%;
